change all bytes to represent intf to int (Comment: #8788)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@29044 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ uvm/api/com/untangle/uvm/shield/ShieldRejectionEvent.java
@@ -95,7 +95,7 @@ import com.untangle.uvm.logging.SyslogPriority;
             return this.clientIntf;
         }
 
-        public void setClientIntf( byte clientIntf )
+        public void setClientIntf( int clientIntf )
         {
             this.clientIntf = clientIntf;
         }
